增加删除弹窗

+++ src/views/strategyGroup.vue
@@ -609,25 +609,36 @@
           strategyGroupName: rowData.strategyGroupName,
           operatorId: rowData.operatorId
         }
-        Api.deleteStrategygroup(params).then((res) => {
-          if(res.code === 200) {
+        this.$confirm('确定删除该策略组吗?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+          Api.deleteStrategygroup(params).then((res) => {
+            if(res.code === 200) {
+              this.$message({
+                message: "删除策略组成功",
+                type: 'success'
+              })
+            } else {
+              this.$message({
+                message: res.msg,
+                type: 'warning'
+              })
+            }
+          }).catch((err => {
+            console.log('err', err)
             this.$message({
-              message: "删除策略组成功",
-              type: 'success'
-            })
-          } else {
-            this.$message({
-              message: res.msg,
+              message: "删除策略组失败",
               type: 'warning'
             })
-          }
-        }).catch((err => {
-          console.log('err', err)
+          }))
+        }).catch(() => {
           this.$message({
-            message: "删除策略组失败",
-            type: 'warning'
-          })
-        }))
+            type: 'info',
+            message: '已取消删除'
+          });
+        });
       },
       // 分页查询策略组
       strategygroupListpage(page, pageSize) {
